% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apaebmet.R
\docType{data}
\name{apaebmet}
\alias{apaebmet}
\title{Meteorological Data from Apalachicola Bay - East Bay Station}
\format{A \code{\link[base]{data.frame}} object}
\source{
\href{http://cdmo.baruch.sc.edu/}{CDMO}
}
\usage{
data(apaebmet)
}
\description{
Meteorological data from Apalachicola Bay - East Bay station
}
\references{
NOAA National Estuarine Research Reserve System (NERRS). System-wide Monitoring Program. Data accessed from the NOAA NERRS Centralized Data Management Office website: http://www.nerrsdata.org/; accessed 08 October 2016
}
\keyword{datasets}
