% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_season.R
\name{assign_season}
\alias{assign_season}
\title{Assign seasons to SWMP sampling data}
\usage{
assign_season(data, season = NULL, season_names = NULL,
  season_start = NULL, abb = FALSE)
}
\arguments{
\item{data}{a vector of POSIXct dates}

\item{season}{A list of seasons. Months (1-12) are assigned to different groups based on user preference. Defaults to 12 months, starting with January. Must assign a minimum of two seasons}

\item{season_names}{A string vector of season names. The number of season names must match the length of the \code{season} list. A minimum of two seasons must be assigned (e.g., 'Wet', 'Dry'). Defaults to 12 months, starting with January. The number of season names must match the number of seasons}

\item{season_start}{defaults to 12 months, starting with January}

\item{abb}{logical, should abbreviations for month names be used? Defaults to \code{FALSE}}
}
\value{
Returns a vector of ordered 
season factors.
}
\description{
Assign seasons to SWMPr sampling data on a monthly basis or user-defined basis
}
\details{
A helper function used by several data analyses to assign seasons to sampling data and to order the seasons. This function also allows the user to designate which season should be first.
}
\examples{
\dontrun{
data(apacpwq)
dat <- apacpwq

seas <- assign_season(dat$datetimestamp)
levels(seas)

seas <- assign_season(dat$datetimestamp, abb = TRUE)
levels(seas)

seas <- assign_season(dat$datetimestamp, season_start = 'March')
levels(seas)

seas <- assign_season(dat$datetimestamp, 
season = list(c(1,2,3), c(4,5,6), c(7,8,9), c(10, 11, 12)),
season_names = c('Winter', 'Spring', 'Summer', 'Fall'), season_start = 'Spring')
levels(seas)

seas <- assign_season(dat$datetimestamp, season = list(c(10:12, 1:3), c(4:9)), 
season_names = c('Wet', 'Dry'))
levels(seas)
}

}
