% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_map.R
\name{base_map}
\alias{base_map}
\title{Create background map}
\usage{
base_map(bbox, bg_epsg, ...)
}
\arguments{
\item{bbox}{Bounding box vector.}

\item{bg_epsg}{EPSG code for}

\item{...}{Additional arguments to be passed to \code{\link[ggmap::get_stamenmap]}}
}
\value{
returns a \code{\link[ggplot2]} object
}
\description{
Create background map based on raster map retrieved with \code{\link[ggmap::get_stamenmap]}.  If \code{\link[ggmap]} is unavailable, creates simple map using OpenStreetMap vector layers.
}
\details{
A helper function to create background map based on OpenStreetMap vector layers.
}
\author{
Dave Eslinger
}
\concept{miscellaneous}
