% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbm_spatial.R
\docType{data}
\name{cbm_spatial}
\alias{cbm_spatial}
\title{Spatial Data from Chesapeake Bay - Maryland}
\format{A \code{\link[sp]{SpatialPolygons}} object}
\source{
\href{http://cdmo.baruch.sc.edu/get/gis.cfm}{CDMO}
}
\usage{
data(cbm_spatial)
}
\description{
Shapefile for Chesapeake Bay - Maryland reserve boundary
}
\references{
NOAA National Estuarine Research Reserve System (NERRS). System-wide Monitoring Program. Data accessed from the NOAA NERRS Centralized Data Management Office website: http://www.nerrsdata.org/; accessed 08 October 2016
}
\keyword{datasets}
