% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counties_4269.R
\docType{data}
\name{counties_4269}
\alias{counties_4269}
\title{US County Map}
\format{
A \code{sf]{DataFrame}} object
}
\source{
\href{https://www.census.gov/geographies/mapping-files.html}{US Census Bureau}
}
\usage{
data('counties_4269')
}
\description{
US County boundaries from the US Census Bureau's MAF/TIGER geographic database in EPSG:4269, with all data except state FIPS codes and land area removed.
}
\references{
United States Census Bureau. Data accessed from the US Census Bureau website: https://www2.census.gov/geo/tiger/GENZ2018/shp/cb_2018_us_county_20m.zip; accessed 13 March 2020
}
\keyword{datasets}
