% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elk_spatial.R
\docType{data}
\name{elk_spatial}
\alias{elk_spatial}
\title{Spatial Data from Elkhorn Slough}
\format{
A \code{\link[sf]{sf}} object
}
\source{
\href{https://cdmo.baruch.sc.edu/get/gis.cfm}{CDMO}
}
\usage{
data(elk_spatial)
}
\description{
Shapefile for Elkhorn Slough reserve boundary
}
\references{
NOAA National Estuarine Research Reserve System (NERRS). System-wide Monitoring Program. Data accessed from the NOAA NERRS Centralized Data Management Office website: https://www.nerrsdata.org/; accessed 08 October 2016
}
\keyword{datasets}
