% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elksmnut.R
\docType{data}
\name{elksmnut}
\alias{elksmnut}
\title{Nutrient Data from Elkhorn Slough - South Marsh Station}
\format{A \code{\link[base]{data.frame}} object}
\source{
\href{http://cdmo.baruch.sc.edu/}{CDMO}
}
\usage{
data(elksmnut)
}
\description{
Nutrient data from Elkhorn Slough South Marsh station
}
\references{
NOAA National Estuarine Research Reserve System (NERRS). System-wide Monitoring Program. Data accessed from the NOAA NERRS Centralized Data Management Office website: http://www.nerrsdata.org/; accessed 08 October 2016
}
\keyword{datasets}
