% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft_col_names.R
\name{ft_col_names}
\alias{ft_col_names}
\title{Convert Parameter Abbreviations}
\usage{
ft_col_names(param)
}
\arguments{
\item{param}{chr, vector of parameter abbreviations}
}
\value{
Returns a \code{data.frame} of user-specified results to be displayed
}
\description{
Convert \code{SWMPr} parameter abbreviations into formats appropriate for use with NERRS reserve level template \code{\link[flextable]{flextable}}
}
\details{
A helper function used internally by \code{\link{create_sk_flextable_list}} to label \code{\link[flextable]{flextable}} columns in the trend table for the reserve level report.
}
\author{
Julie Padilla
}
\concept{reporting}
