% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geographic_unique_stations.R
\name{geographic_unique_stations}
\alias{geographic_unique_stations}
\title{Return a vector of geographically unique NERR Stations}
\usage{
geographic_unique_stations(nerr_site_id)
}
\arguments{
\item{nerr_site_id}{chr vector of valid NERR stations}
}
\value{
returns a vector of NERR stations
}
\description{
Creates a alphabetically sorted, vector of geographically unique stations for mapping
}
\details{
This function takes a vector of NERR stations and only returns geographically unique locations. Intended for use with \code{\link{res_local_map}}
}
\examples{
\dontrun{
stns <- c('apacpnut', 'apacpwq', 'apadbnut', 'apadbwq', 'apaebmet',
'apaebnut', 'apaebwq', 'apaesnut', 'apaeswq')

geographic_unique_stations(stns)
}

}
\author{
Julie Padilla
}
\concept{
mapping
}
