% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_shp_name.R
\name{get_shp_name}
\alias{get_shp_name}
\title{Identify shapefile for NERRS reserve}
\usage{
get_shp_name(gis.file.loc)
}
\arguments{
\item{gis.file.loc}{path to gis file location}
}
\value{
Returns a character string of the shapefile for the reserve boundary
}
\description{
Identify the shapefile name associated with the reserve in the data file
}
\details{
This function is intended for internal use with the NERRS reserve level reporting scripts. It identifies the name of the shapefile associated with the NERRS reserve.
}
\author{
Julie Padilla
}
\concept{reporting}
