% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_site_code.R
\name{get_site_code}
\alias{get_site_code}
\title{Return NERRS reserve site code based on data in the data file}
\usage{
get_site_code(data.file)
}
\arguments{
\item{data.file}{data source location}
}
\value{
Returns reserve site code as chr
}
\description{
A helper function for the NERRS automated reporting template
}
\concept{
reporting
}
