% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_site_coordinates.R
\name{get_site_coordinates}
\alias{get_site_coordinates}
\title{Identify NERRS reserve from metadata}
\usage{
get_site_coordinates(data.file, active = TRUE)
}
\arguments{
\item{data.file}{location of data}

\item{active}{logical. Only return active stations?}
}
\value{
Returns a dataframe of station ids, station names, lat/long
}
\description{
Identify the NERRS reserve from metadata in the data file
}
\concept{
reporting
}
