% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historical_daily_range.R
\name{historical_daily_range}
\alias{historical_daily_range}
\alias{historical_daily_range.swmpr}
\title{Historical Daily Range Timeseries}
\usage{
historical_daily_range(swmpr_in, ...)

\method{historical_daily_range}{swmpr}(swmpr_in, param = NULL,
  hist_rng = NULL, target_yr = NULL, criteria = NULL, log_trans = FALSE,
  plot_title = FALSE, plot = TRUE, ...)
}
\arguments{
\item{swmpr_in}{input swmp object}

\item{...}{additional arguments passed to other methods (not used for this function).}

\item{param}{chr string of variable to plot}

\item{hist_rng}{numeric vector, if historic range is not specified then the min/max values of the data set will be used.}

\item{target_yr}{numeric, the target year that should be compared against the historic range. If target year is not specified then dot will not be plotted}

\item{criteria}{numeric, a numeric criteria that will be plotted as a horizontal line}

\item{log_trans}{logical, should y-axis be log? Defaults to \code{FALSE}}

\item{plot_title}{logical, should the station name be included as the plot title? Defaults to \code{FALSE}}

\item{plot}{logical, should a plot be returned? Defaults to \code{TRUE}}
}
\value{
A \code{\link[ggplot2]{ggplot}} object
}
\description{
Daily averages and variability. Looking at variability within a year, with historical context
}
\details{
Comparison of analysis year statistical summaries to the long term monthly averages and variability.
To put it another way, this analysis is looking at monthly averages in context of comparison to period variability.
}
\examples{
\dontrun{
data(apacpwq)
dat <- apacpwq

dat <- qaqc(apacpwq, qaqc_keep = c('0', '3', '5'))
# with criteria
y <- historical_daily_range(dat, param = 'do_mgl', target_yr = 2013, criteria = 2, abb = T)

# w/o criteria
x <- historical_daily_range(dat, param = 'do_mgl', target_yr = 2013, abb = T)

# add a y label
x <- x + labs(x = NULL, y = "Dissolved Oxygen (mg/L)")
}
}
\seealso{
\code{\link[ggplot2]{ggplot}}, \code{\link{assign_season}}
}
\author{
Julie Padilla
}
\concept{
analyze

analyze

analyze
}
