% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historical_range.R
\name{historical_range}
\alias{historical_range}
\alias{historical_range.swmpr}
\title{Historical Range Timeseries}
\usage{
historical_range(swmpr_in, ...)

\method{historical_range}{swmpr}(swmpr_in, param = NULL, hist_rng = NULL,
  target_yr = NULL, criteria = NULL, log_trans = FALSE, plot = TRUE,
  ...)
}
\arguments{
\item{swmpr_in}{input swmpe object}

\item{...}{additional arguments passed to other methods. See \code{\link{assign_season}}}

\item{param}{chr string of variable to plot}

\item{hist_rng}{numeric vector, if historic range is not specified then the min/max values of the data set will be used.}

\item{target_yr}{numeric, the target year that should be compared against the historic range. If target year is not specified then dot will not be plotted}

\item{criteria}{numeric, a numeric criteria that will be plotted as a horizontal line}

\item{log_trans}{logical, should y-axis be log? Defaults to \code{FALSE}}

\item{plot}{logical, should a plot be returned? Defaults to \code{TRUE}}
}
\value{
A \code{\link[ggplot2]{ggplot}} object
}
\description{
Monthly average and variability. Looking at variability within each month; no historical context
}
\details{
Comparison of analysis year statistical summaries to the long term monthly averages and variability.
To put it another way, this analysis is looking at monthly averages in context of comparison to period variability.
}
\examples{
\dontrun{
data(apacpwq)
dat <- apacpwq

dat <- qaqc(apacpwq, qaqc_keep = c('0', '3', '5'))
do_plt <- historical_range(dat, param = 'do_mgl', target_yr = 2012)
do_plt <- historical_range(dat, param = 'do_mgl', target_yr = 2012, criteria = 2)
}
}
\seealso{
\code{\link[ggplot2]{ggplot}}, \code{\link{assign_season}}
}
\author{
Julie Padilla
}
\concept{
analyze

analyze

analyze
}
