% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_p_labs.R
\name{lm_p_labs}
\alias{lm_p_labs}
\title{P-Value labels for Plotting}
\usage{
lm_p_labs(dat_in)
}
\arguments{
\item{dat_in}{\code{data.frame} with year, season, min, mean, max columns}
}
\value{
Returns \code{data.frame} for use with \code{seasonal_dot}
}
\description{
Generate a dataframe of p-value labels based on p-values from linear regression
}
\details{
This function returns a data.frame of p-value labels for use with the seasonal dot plot. P-values are taken from linear regression.
}
\seealso{
\code{lm}
}
