% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_shp_file.R
\name{load_shp_file}
\alias{load_shp_file}
\title{Load and format shapefile for reserve level map}
\usage{
load_shp_file(path, dissolve_boundaries = T)
}
\arguments{
\item{path}{path to shapefile and name}

\item{dissolve_boundaries}{logical, should reserve boundaries be dissolved? Defaults to \code{TRUE}}
}
\value{
Returns a character string of the shapefile for the reserve boundary
}
\description{
Load shapefile for use with res_local_map
}
\details{
Creates a base map of the US with options for including AK, HI, and PR. The user can also
This function was developed from a blog post by Bob Rudis (https://rud.is/b/2014/11/16/moving-the-earth-well-alaska-hawaii-with-r/)
}
\author{
Julie Padilla
}
\concept{
reporting
}
