% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_boxplot.R
\name{raw_boxplot}
\alias{raw_boxplot}
\alias{raw_boxplot.swmpr}
\title{Seasonal boxplots of raw data}
\usage{
raw_boxplot(swmpr_in, ...)

\method{raw_boxplot}{swmpr}(swmpr_in, param = NULL, target_yr = NULL,
  criteria = NULL, log_trans = FALSE, plot_title = FALSE, ...)
}
\arguments{
\item{swmpr_in}{input swmp object}

\item{...}{additional arguments passed to other methods. See \code{\link{assign_season}} and \code{\link{y_labeler}}.}

\item{param}{chr string of variable to plot}

\item{target_yr}{numeric, if target year is not specified then dot will not be plotted. If target year is not specified the most recent year in the \code{swmpr_in} will be used.}

\item{criteria}{numeric, a numeric criteria that will be plotted as a horizontal line}

\item{log_trans}{logical, should y-axis be log? Defaults to \code{FALSE}}

\item{plot_title}{logical, should the station name be included as the plot title? Defaults to \code{FALSE}}
}
\value{
A \code{\link[ggplot2]{ggplot}} object
}
\description{
Boxplots of raw data by user-defined season
}
\details{
Annual time series for year of interest
}
\seealso{
\code{\link[ggplot2]{ggplot}}, \code{\link{assign_season}}
}
\author{
Julie Padilla
}
\concept{
analyze

analyze
}
