% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_inf_and_nan.R
\name{remove_inf_and_nan}
\alias{remove_inf_and_nan}
\title{Replace Inf/-Inf/NaN values}
\usage{
remove_inf_and_nan(x)
}
\arguments{
\item{x}{input matrix}
}
\value{
Returns a \code{matrix} object
}
\description{
Replace \code{Inf}, \code{-Inf}, and \code{NaN} in a matrix with \code{NA}
}
\details{
This function replaces \code{Inf}, \code{-Inf}, and \code{NaN} in a matrix with \code{NA}. It is used internally by several functions.
}
\author{
Julie Padilla
}
\concept{miscellaneous}
