% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res_local_map.R
\name{res_local_map}
\alias{res_local_map}
\title{Local Reserve Map}
\usage{
res_local_map(nerr_site_id, stations, bbox)
}
\arguments{
\item{nerr_site_id}{chr string of the reserve to make, first three characters used by NERRS}

\item{stations}{chr string of the reserve stations to include in the map}

\item{bbox}{a bounding box associated with the reserve}
}
\description{
Create a stylized reserve level map for use with the reserve level reporting template
}
\details{
Creates a base map of the US with options for including AK, HI, and PR. The user can also
This function was developed from a blog post by Bob Rudis (https://rud.is/b/2014/11/16/moving-the-earth-well-alaska-hawaii-with-r/)
}
\examples{
\dontrun{

##Just the national map

## a compact reserve

## a multicomponent reserve
}
}
