% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res_national_map.R
\name{res_national_map}
\alias{res_national_map}
\title{Reserve National Map}
\usage{
res_national_map(incl = c("contig", "AK", "HI", "PR"),
  highlight_states = NULL, highlight_reserves = NULL, agg_county = T)
}
\arguments{
\item{incl}{Str vector to include AK, HI , and PR (case sensitive)}

\item{highlight_states}{character vector of state FIPS codes}

\item{highlight_reserves}{character vector of 3 letter reserve codes}

\item{agg_county}{logical, should counties be aggregated tot he state-level? Defaults to \code{TRUE}}
}
\value{
Returns a \code{\link[ggplot2]{ggplot}} object
}
\description{
Create a base map for NERRS reserves in ggplot
}
\details{
Creates a base map of the US with options for including AK, HI, and PR. The user can also
This function was developed from a blog post by Bob Rudis (https://rud.is/b/2014/11/16/moving-the-earth-well-alaska-hawaii-with-r/)
}
\examples{
\dontrun{

##Just the national map
res_national_map()

##National map highlighting states with NERRS
nerr_states <- c('01', '02', '06', '10', '12', '13', '15'
, '23', '24', '25', '27', '28', '33', '34', '36', '37', '39'
, '41', '44', '45', '48', '51', '53', '55', '72')

res_national_map(highlight_states = nerr_states)

##National map highlighting west coast states and NERRS (including AK)
nerr_states_west <- c('02', '06', '41', '53')

nerrs_codes <- c('pdb', 'sos', 'sfb', 'elk', 'tjr', 'kac')

res_national_map(highlight_states = nerr_states_west, highlight_reserve = nerrs_codes)
}

}
\author{
Bob Rudis, Julie Padilla
}
\concept{
mapping
}
