% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_stations.R
\docType{data}
\name{sampling_stations}
\alias{sampling_stations}
\title{Detailed of NERRS site data}
\format{
A \code{\link[base]{data.frame}} object
}
\source{
\href{http://cdmo.baruch.sc.edu/}{CDMO}
}
\usage{
data(sampling_stations)
}
\description{
Metadata on NERRS stations provided by the Central Data Management Office (CDMO) when data is downloaded
}
\references{
NOAA National Estuarine Research Reserve System (NERRS). System-wide Monitoring Program. Data accessed from the NOAA NERRS Centralized Data Management Office website: http://www.nerrsdata.org/; accessed 08 October 2016
}
\keyword{datasets}
