% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_stations_backup.R
\docType{data}
\name{sampling_stations_backup}
\alias{sampling_stations_backup}
\title{A Backup of Detailed NERRS site data}
\format{
A \code{\link[base]{data.frame}} object
}
\source{
\href{http://cdmo.baruch.sc.edu/}{CDMO}
}
\usage{
data(sampling_stations_backup)
}
\description{
A backup of the current metadata on NERRS stations provided by the Central Data Management Office (CDMO) created when new data is loaded
}
\references{
NOAA National Estuarine Research Reserve System (NERRS). System-wide Monitoring Program. Data accessed from the NOAA NERRS Centralized Data Management Office website: http://www.nerrsdata.org/; accessed 08 October 2016
}
\keyword{datasets}
