% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal_boxplot.R
\name{seasonal_boxplot}
\alias{seasonal_boxplot}
\title{Seasonal boxplots}
\usage{
seasonal_boxplot(swmpr_in, ...)
}
\arguments{
\item{swmpr_in}{input swmpe object}

\item{...}{additional arguments passed to other methods. See \code{\link{assign_season}}}

\item{param}{chr string of variable to plot}

\item{target_yr}{numeric, if target year is not specified then dot will not be plotted}

\item{log_trans}{logical, should y-axis be log? Defaults to \code{FALSE}}

\item{criteria}{numeric, a numeric criteria that will be plotted as a horizontal line}

\item{FUN}{function used to aggregate daily SWMP data}
}
\value{
A \code{\link[ggplot2]{ggplot}} object
}
\description{
Annual time series for year of interest on top of long-term percentiles
}
\details{
Annual time series for year of interest on top of long-term percentiles
}
\seealso{
\code{\link[ggplot2]{ggplot}}
}
\concept{
analyze
}
