% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_date_breaks.R
\name{set_date_break_labs}
\alias{set_date_break_labs}
\title{Set reasonable date breaks labels}
\usage{
set_date_break_labs(rng)
}
\arguments{
\item{rng}{date range years}
}
\description{
Select reasonable labels for breaks used in scale_x_datetime
}
\details{
Annual time series for year of interest on top of long-term percentiles
}
\author{
Julie Padilla
}
\concept{
analyze
}
