% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_date_breaks.R
\name{set_date_breaks_minor}
\alias{set_date_breaks_minor}
\title{Select reasonable minor breaks for \code{\link[ggplot2]{scale_x_datetime}}}
\usage{
set_date_breaks_minor(rng)
}
\arguments{
\item{rng}{date range years}
}
\value{
Returns a chr string for \code{date_breaks}
}
\description{
Select reasonable minor breaks for \code{\link[ggplot2]{scale_x_datetime}}
}
\details{
A helper function for easier date label setting
}
\seealso{
\code{\link{set_date_break_labs}}, \code{\link[ggplot2]{scale_x_datetime}}
}
\author{
Dave Eslinger, Julie Padilla
}
\concept{miscellaneous}
