% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sk_monthly.R
\name{sk_monthly}
\alias{sk_monthly}
\alias{sk_monthly.swmpr}
\title{Seasonal Kendall Analysis for Monthly data}
\usage{
sk_monthly(swmpr_in, ...)

\method{sk_monthly}{swmpr}(swmpr_in, param = NULL, stat_lab = "Average",
  FUN = function(x) mean(x, na.rm = T), ...)
}
\arguments{
\item{swmpr_in}{input swmpr object}

\item{...}{additional arguments passed to other methods. See \code{\link{assign_season}}}

\item{param}{chr string of variable to plot}

\item{stat_lab}{chr, label for the summary statistic defined in \code{FUN}. Defaults to "Average"}

\item{FUN}{function used to aggregate monthly SWMP data}
}
\value{
A \code{data.frame} object
}
\description{
Seasonal trends
}
\details{
Seasonal Kendall
}
\examples{
\dontrun{

dat_wq <- elksmwq
dat_wq <- subset(dat_wq, subset = c('2007-01-01 0:00', '2017-01-01 0:00'))
dat_wq <- qaqc(dat_wq, qaqc_keep = c(0, 3, 5))
}
}
\seealso{
\code{\link[ggplot2]{ggplot}}, \code{\link[EnvStats]{kendallSeasonalTrendTest}}
}
\author{
Julie Padilla
}
\concept{
analyze

analyze
}
