% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_param_check.R
\name{std_param_check}
\alias{std_param_check}
\title{Standard Parameter Check}
\usage{
std_param_check(param)
}
\arguments{
\item{param}{chr string of variable abbreviation}
}
\value{
Returns \code{TRUE} or \code{FALSE}
}
\description{
Determine if a parameter is one of the standard SWMP parameters
}
\details{
A helper function used internally by several plotting functions to determine if parameter has a standard y-axis label.
}
