% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/title_labeler.R
\name{title_labeler}
\alias{title_labeler}
\title{Generate Plot Title Based on NERR Site ID}
\usage{
title_labeler(nerr_site_id)
}
\arguments{
\item{nerr_site_id}{chr string of NERR site id}
}
\value{
Returns character vector
}
\description{
Generate a plot title based on SWMP station abbreviation
}
\details{
A helper function used internally by several plotting functions to generate plot titles.
}
\examples{
ttl <- title_labeler('elkapwq')

}
\author{
Julie Padilla
}
\concept{miscellaneous}
