% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/us_aea.R
\docType{data}
\name{us_aea}
\alias{us_aea}
\title{US County Map}
\format{A \code{\link[sp]{SpatialPolygonsDataFrame}} object}
\source{
\href{https://www.census.gov/geo/maps-data/data/cbf/cbf_counties.html}{US Census Bureau}
}
\usage{
data(us_aea)
}
\description{
US County boundaries from the US Census Bureau's MAF/TIGER geographic database. Reprojected using Albers Equal Area.
}
\references{
United States Census Bureau. Data accessed from the US Census Bureau website: http://www2.census.gov/geo/tiger/GENZ2010/gz_2010_us_050_00_20m.zip; accessed 06 April 2018
}
\keyword{datasets}
