% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/us_laea.R
\docType{data}
\name{us_laea}
\alias{us_laea}
\title{US County Map}
\format{
A \code{\link[sp]{SpatialPolygonsDataFrame}} object
}
\source{
\href{https://www.census.gov/geographies/mapping-files.html}{US Census Bureau}
}
\usage{
data(us_laea)
}
\description{
US County boundaries from the US Census Bureau's MAF/TIGER geographic database. Reprojected using Lambert Azimuthal Equal Area.
}
\references{
United States Census Bureau. Data accessed from the US Census Bureau website: http://www2.census.gov/geo/tiger/GENZ2010/gz_2010_us_050_00_20m.zip; accessed 06 April 2018
}
\keyword{datasets}
