% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_map.R
\name{base_map}
\alias{base_map}
\title{Create background map}
\usage{
base_map(bbox, bg_crs = 4326, ...)
}
\arguments{
\item{bbox}{Bounding box vector.}

\item{...}{Additional arguments to be passed to ggmap::get_stamenmap}

\item{bg_epsg}{EPSG code for returned map.}
}
\value{
Returns a \code{\link[ggplot2]{ggplot}} object#'
}
\description{
Create background map
}
\details{
A helper function to create background map based on based on raster map tiles retrieved with \code{\link[ggmap::get_stamenmap]}ggmap::get_stamenmap.  If \code{\link[ggmap]}ggmap is unavailable, creates simple map using county-level polygon files.  This will be a pretty coarse resolution map and should be considered a placeholder.
}
\author{
Dave Eslinger
}
\concept{miscellaneous}
