% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_map.R
\name{base_map}
\alias{base_map}
\title{Create background map}
\usage{
base_map(bbox, bg_crs = 4326, maptype = "toner-lite", zoom = NULL, ...)
}
\arguments{
\item{bbox}{Bounding box vector.}

\item{bg_crs}{EPSG code or \code{st_crs} object for the returned map.}

\item{maptype}{Background map type from Stamen Maps (\url{http://maps.stamen.com/}); one of c("terrain", "terrain-background", "terrain-labels", "terrain-lines", "toner", "toner-2010", "toner-2011", "toner-background", "toner-hybrid", "toner-labels", "toner-lines", "toner-lite", "watercolor").}

\item{zoom}{Zoom level for the base map created when \code{bg_map} is not specified.  An integer value, 5 - 15, with higher numbers providing  more detail.  If not provided, a zoom level is autoscaled based on \code{bbox} parameters.}

\item{...}{Additional arguments to be passed to \code{ggmap::get_stamenmap}}
}
\value{
Returns a \code{ggplot2} object.
}
\description{
Create background map from a bounding box
}
\details{
A helper, or stand-alone, function to create background map based on based on raster map tiles retrieved with \code{ggmap::get_stamenmap}.  If \code{ggmap} is unavailable, the function creates a basic map using county-level polygon files.  This map is fairly crude and should be considered a placeholder.
}
\examples{

# Simple, low-zoom  map for testing
bound_box <- c(-77.393, 39.741, -75.553, 38.277)
x <- base_map(bound_box, zoom = 10)

\donttest{
# Default zoom map with watercolor maptype.
y <- base_map(bound_box, maptype = 'watercolor')
}
}
\author{
Dave Eslinger
}
\concept{analyze}
