% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_map.R
\name{base_map}
\alias{base_map}
\title{Create background map}
\usage{
base_map(bbox, bg_crs = 4326, ...)
}
\arguments{
\item{bbox}{Bounding box vector.}

\item{bg_crs}{EPSG code for returned map.}

\item{...}{Additional arguments to be passed to \code{ggmap::get_stamenmap}}
}
\value{
Returns a \code{\link[ggplot2]{ggplot}} object.
}
\description{
Create background map
}
\details{
A helper function to create background map based on based on raster map tiles retrieved with \code{ggmap::get_stamenmap}.  If \code{ggmap} is unavailable, the function creates a basic map using county-level polygon files.  This map is fairly crude and should be considered a placeholder.
}
\author{
Dave Eslinger
}
\concept{miscellaneous}
