% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_map.R
\name{base_map}
\alias{base_map}
\title{Create background map}
\usage{
base_map(bbox, bg_crs)
}
\arguments{
\item{bbox}{Bounding box vector,}
}
\value{
returns a \code{\link[ggplot2]} object
}
\description{
Create background map based on OpenStreetMap vector layers.
}
\details{
A helper function to create background map based on OpenStreetMap vector layers.
}
\author{
Dave Eslinger
}
\concept{miscellaneous}
