% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reserve.R
\name{get_reserve}
\alias{get_reserve}
\title{Identify NERRS reserve from metadata}
\usage{
get_reserve(data.file)
}
\arguments{
\item{data.file}{location of data}
}
\value{
Returns a character string of the full reserve name
}
\description{
Identify the NERRS reserve from metadata in the data file
}
\details{
This function is intended for internal use with the NERRS reserve level reporting scripts. It determines the name of the full name of the NERRS reserve associated with the data in the user-specified data folder.
}
\author{
Julie Padilla
}
\concept{reporting}
