% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_site_code.R
\name{get_site_code}
\alias{get_site_code}
\title{Return NERRS reserve site code based on data in the data file}
\usage{
get_site_code(data.file)
}
\arguments{
\item{data.file}{data source location}
}
\value{
Returns 3-letter, reserve site code as chr
}
\description{
Identify the 3-letter NERRS reserve code from metadata in the data file
}
\details{
This function is intended for internal use with the NERRS reserve level reporting scripts. It returns the 3-letter reserve code associated with the data in the user-specified data folder.
}
\author{
Julie Padilla
}
\concept{reporting}
