% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_site_coordinates.R
\name{get_site_coordinates}
\alias{get_site_coordinates}
\title{Identify NERRS sampling locations from metadata}
\usage{
get_site_coordinates(data.file, active = TRUE)
}
\arguments{
\item{data.file}{location of data}

\item{active}{logical. Only return active stations?}
}
\value{
Returns a dataframe of station ids, station names, lat/long
}
\description{
Identify the latitude/longitude for sampling stations based on the metadata in the data file
}
\details{
This function is intended for internal use with the NERRS reserve level reporting scripts. It returns the names, station codes, and coordinates associated with the data in the user-specified data folder.
}
\author{
Julie Padilla
}
\concept{reporting}
