% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sites.R
\name{get_sites}
\alias{get_sites}
\title{Identify NERRS reserve stations from metadata}
\usage{
get_sites(data.file, type = c("wq", "nut", "met"), active = TRUE)
}
\arguments{
\item{data.file}{location of data}

\item{type}{chr string of data station type (\code{'wq'}, \code{'nut'}, or \code{'met'})}

\item{active}{logical. Should inactive stations be excluded?}
}
\value{
Returns a character vector of reserve stations
}
\description{
Identify the NERRS reserve sampling stations based on the metadata in the data file
}
\details{
This function returns the sampling stations associated with the data in the user-specified data file
}
\concept{
reporting

reporting
}
