% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_p_labs.R
\name{lm_p_labs}
\alias{lm_p_labs}
\title{P-Value labels for Plotting}
\usage{
lm_p_labs(dat_in)
}
\arguments{
\item{dat_in}{\code{data.frame} with year, season, min, mean, max columns}
}
\value{
Returns \code{data.frame} for use with \code{\link{seasonal_dot}}
}
\description{
Generate a dataframe of p-value labels based on p-values from linear regression
}
\details{
A helper function that returns a \code{data.frame} of p-value labels for use with the \code{\link{seasonal_dot}}. P-values are taken from linear regression \code{lm}.
}
\seealso{
\code{\link[stats]{lm}}
}
\author{
Julie Padilla, Dave Eslinger
}
\concept{miscellaneous}
