% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/national_sk_map.R
\name{national_sk_map}
\alias{national_sk_map}
\title{Reserve National Map with Seasonal Kendall Results}
\usage{
national_sk_map(incl = c("contig", "AK", "HI", "PR"),
  highlight_states = NULL, sk_reserves = NULL, sk_results = NULL,
  sk_fill_colors = c("#247BA0", "#A3DFFF", "#595959", "#595959"),
  agg_county = T, color_border = "#5b5b5b", fill_highlight = "#f8f8f8",
  fill_states = "b7b7b7")
}
\arguments{
\item{incl}{chr vector to include AK, HI , and PR (case sensitive)}

\item{highlight_states}{chr vector of state FIPS codes}

\item{sk_reserves}{chr vector of 3 letter reserve codes that have seasonal kendall results}

\item{sk_results}{chr vector of seasonal kendall results. Results can be 'inc', 'dec', 'insig', or 'insuff' which stand for 'increasing trend', 'decreasing trend', 'statistically insignificant trend', or 'insufficient data to detect trend'}

\item{sk_fill_colors}{chr vector of colors used to fill seasonal kendall result markers}

\item{agg_county}{logical, should counties be aggregated tot he state-level? Defaults to \code{TRUE}}

\item{color_border}{chr, assign color for the boundaries between states}

\item{color_highlight}{chr, assign color for the states listed in \code{highlight_states}}

\item{color}{chr, assign color of states that are not listed in \code{highlight_states}}
}
\value{
Returns a \code{\link[ggplot2]{ggplot}} object
}
\description{
Create a base map for NERRS reserves in ggplot with seasonal kendall results
}
\details{
Create a base map of the US with options for including AK, HI, and PR. The user can choose which states and NERRS reserves to highlight.
This function was developed, in part, from a blog post by Bob Rudis.
}
\examples{
\dontrun{

##Just the national map
res_national_map()

##National map highlighting states with NERRS
nerr_states <- c('01', '02', '06', '10', '12', '13', '15'
, '23', '24', '25', '27', '28', '33', '34', '36', '37', '39'
, '41', '44', '45', '48', '51', '53', '55', '72')

res_national_map(highlight_states = nerr_states)

##National map highlighting west coast states and NERRS (including AK)
nerr_states_west <- c('02', '06', '41', '53')

nerrs_codes <- c('pdb', 'sos', 'sfb', 'elk', 'tjr')
nerrs_sk_results <- c('inc', 'inc', 'dec', 'insig', 'insuff')

national_sk_map(highlight_states = nerr_states_west,
highlight_reserve = nerrs_codes, sk_results = nerrs_sk_results)
}

}
\references{
Rudis, Bob. 2014. "Moving The Earth (well, Alaska & Hawaii) With R". rud.is (blog). November 16, 2014. https://rud.is/b/2014/11/16/moving-the-earth-well-alaska-hawaii-with-r/
}
\author{
Bob Rudis, Julie Padilla
}
\concept{
analyze
}
