% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reserve_locs.R
\name{reserve_locs}
\alias{reserve_locs}
\title{Reserve National Map}
\usage{
reserve_locs(incl = c("contig", "AK", "HI", "PR"), subset_reserve = NULL)
}
\arguments{
\item{incl}{Str vector to include AK, HI , and PR}

\item{subset_reserve}{chr string of 3 letter reserve IDs to include as data points. To include He'eia use the reserve code 'HEA'}
}
\value{
Returns a \code{dataframe} for internal use with \code{res_national_map}
}
\description{
Create a dataframe of selected NERRS locations
}
\details{
A helper function used to generate a dataframe of selected reserve locations for use with \code{res_national_map}
}
\author{
Bob Rudis, Julie Padilla
}
\concept{
analyze
}
