% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal_barplot.R
\name{seasonal_barplot}
\alias{seasonal_barplot}
\alias{seasonal_barplot.swmpr}
\title{Cumulative Bar Plot}
\usage{
seasonal_barplot(swmpr_in, ...)

\method{seasonal_barplot}{swmpr}(swmpr_in, param = NULL, hist_rng = NULL,
  rng_avg = FALSE, log_trans = FALSE, convert = FALSE, plot = TRUE, ...)
}
\arguments{
\item{swmpr_in}{input swmpe object}

\item{...}{additional arguments passed to other methods. See \code{\link{assign_season}}}

\item{param}{chr string of variable to plot}

\item{hist_rng}{numeric vector, if historic range is not specified then the min/max values of the data set will be used.}

\item{rng_avg}{logical, should a longterm average be included on the plot? Defaults to \code{FALSE}}

\item{log_trans}{logical, should y-axis be log? Defaults to \code{FALSE}}

\item{convert}{logical, convert from metric to US units? Defaults to \code{FALSE}}

\item{plot}{logical, should a plot be returned? Defaults to \code{TRUE}}
}
\value{
A \code{\link[ggplot2]{ggplot}} object
}
\description{
Cumulative bar plot over a historic range
}
\details{
Calculated annual bar plot
}
\examples{
\dontrun{
data(apaebmet)
dat <- apaebmet

dat <- qaqc(apaebmet, qaqc_keep = c('0', '3', '5'))
prcp_plt <- seasonal_barplot(dat, param = 'do_mgl', hist_rng = c(2012, 2013))
prcp_plt <- seasonal_barplot(dat, param = 'do_mgl', hist_rng = c(2012, 2013), criteria = 2)
}
}
\seealso{
\code{\link[ggplot2]{ggplot}}, \code{\link{assign_season}}
}
\author{
Julie Padilla
}
\concept{
analyze

analyze

analyze
}
