% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal_boxplot.R
\name{seasonal_boxplot}
\alias{seasonal_boxplot}
\alias{seasonal_boxplot.swmpr}
\title{Seasonal boxplots}
\usage{
seasonal_boxplot(swmpr_in, ...)

\method{seasonal_boxplot}{swmpr}(swmpr_in, param = NULL, hist_rng = NULL,
  target_yr = NULL, criteria = NULL, log_trans = FALSE,
  plot_title = FALSE, plot = T, FUN = function(x) mean(x, na.rm = T), ...)
}
\arguments{
\item{swmpr_in}{input swmpr object}

\item{...}{additional arguments passed to other methods. See \code{\link{assign_season}} and \code{\link{y_labeler}}.}

\item{param}{chr string of variable to plot}

\item{hist_rng}{numeric vector, if historic range is not specified then the min/max values of the data set will be used.}

\item{target_yr}{numeric, the target year that should be compared against the historic range. If target year is not specified then dot will not be plotted}

\item{criteria}{numeric, a numeric criteria that will be plotted as a horizontal line}

\item{log_trans}{logical, should y-axis be log? Defaults to \code{FALSE}}

\item{plot_title}{logical, should the station name be included as the plot title? Defaults to \code{FALSE}}

\item{plot}{logical, should a plot be returned? Defaults to \code{TRUE}}

\item{FUN}{function used to aggregate daily SWMP data}
}
\value{
A \code{\link[ggplot2]{ggplot}} object
}
\description{
Annual time series for year of interest on top of long-term percentiles
}
\details{
Annual time series for year of interest on top of long-term percentiles
}
\seealso{
\code{\link[ggplot2]{ggplot}}
}
\author{
Julie Padilla
}
\concept{
analyze

analyze
}
