% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_date_breaks.R
\name{set_date_break_labs}
\alias{set_date_break_labs}
\title{Set reasonable date breaks labels}
\usage{
set_date_break_labs(rng)
}
\arguments{
\item{rng}{date range years}
}
\value{
Returns a chr string for \code{date_labels}
}
\description{
Select reasonable labels for breaks used in \code{\link[ggplot2]{scale_x_datetime}}
}
\details{
A helper function for easier date label setting
}
\seealso{
\code{\link{set_date_breaks}}, \code{\link[ggplot2]{scale_x_datetime}}
}
\author{
Julie Padilla
}
\concept{analyze}
