% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sk_tidy.R
\name{sk_tidy}
\alias{sk_tidy}
\title{Tidy Seasonal Kendall Results}
\usage{
sk_tidy(data, station, param, stat, alpha = 0.05)
}
\arguments{
\item{data}{a \code{htest} object produced by \code{\link[EnvStats]{kendallSeasonalTrendTest}}}

\item{station}{chr string sampling station}

\item{param}{chr string of variable to plot}

\item{stat}{chr, label to be used for statistic used to group data}

\item{alpha}{num, significance level. Defaults to 0.05}
}
\value{
Returns a \code{data.frame} of results from \code{\link[EnvStats]{kendallSeasonalTrendTest}}
}
\description{
Tidy results from \code{\link[EnvStats]{kendallSeasonalTrendTest}}
}
\details{
A helper function used by \code{\link{sk_seasonal}} to return a table of tidied values.
}
\author{
Julie Padilla
}
\concept{miscellaneous}
