% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_identification.R
\name{threshold_identification}
\alias{threshold_identification}
\alias{threshold_identification.swmpr}
\title{Tabulate Threshold Exceedances}
\usage{
threshold_identification(swmpr_in, ...)

\method{threshold_identification}{swmpr}(swmpr_in, param, parameter_threshold,
  threshold_type, time_threshold = NULL, ...)
}
\arguments{
\item{swmpr_in}{input data object}

\item{...}{arguments passed to other methods}

\item{param}{vector of parameters to evaluate}

\item{parameter_threshold}{vector of numerical thresholds to evaluate parameters against}

\item{threshold_type}{vector of logical operators ('<', '>', '<=', '>=', '==', '!=')}

\item{time_threshold}{The amount of time an event must last to be counted (in hours)}
}
\value{
A data frame of threshold exceedances by parameter
}
\description{
Tabulate user-specified threshold exceedances
}
\details{
Add some details about how this thing works here.
++You must 'setstep' beforehand
++Nutrient thresholds do not require a time threshold, but met and wq do
++Recommended nutrient thresholds are determined by NCCR (say something about DIN)
}
\examples{
\dontrun{
## change the type argument if plotting discrete and continuous data
wq <- apacpwq
nut <- apacpnut

wq_pars<- threshold_identification(dat_wq, param = c('do_mgl', 'ph', 'temp')
                          , parameter_threshold = c('2', '5', '30')
                          , threshold_type = c('<', '<', '>'), time_threshold = 2)
wq_par<- threshold_identification(dat_wq, param = c('do_mgl')
                          , parameter_threshold = c('2')
                          , threshold_type = c('<'), time_threshold = 2)


nut_pars <- threshold_identification(dat_nut, param = c('chla_n', 'po4f')
                          , parameter_threshold = c('10', '0.01')
                          , threshold_type = c('>', '>'))

nut_par <- threshold_identification(dat_nut, param = c('chla_n')
                          , parameter_threshold = c('10')
                          , threshold_type = c('>'))

nut_err <- threshold_identification(dat_nut, param = c('chla_n')
                          , parameter_threshold = c('30')
                          , threshold_type = c('>'))

}
}
\author{
Julie Padilla
}
\concept{
analyze

analyze
}
