% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_percentile_plot.R
\name{threshold_percentile_plot}
\alias{threshold_percentile_plot}
\alias{threshold_percentile_plot.swmpr}
\title{Threshold Percentile Plot}
\usage{
threshold_percentile_plot(swmpr_in, ...)

\method{threshold_percentile_plot}{swmpr}(swmpr_in, param = NULL,
  hist_rng = NULL, target_yr = NULL, percentiles = c(0.05, 0.95),
  by_month = TRUE, log_trans = FALSE, converted = FALSE,
  plot_title = FALSE, plot = TRUE, ...)
}
\arguments{
\item{swmpr_in}{input swmp object}

\item{...}{additional arguments passed to other methods (not used for this function).}

\item{param}{chr string of variable to plot}

\item{hist_rng}{numeric vector of year(s) used to calculate percentiles, if range is not specified then the min/max values of the data set will be used.}

\item{target_yr}{numeric, year of interest for plotting. If not, specified, the entire data set will be plotted.}

\item{percentiles}{numeric vector of percentiles to calculate (maximum: 2). Defaults to 5th and 95th percentiles.}

\item{by_month}{logical. should percentiles be calculated on a monthly basis? Defaults to \code{FALSE}}

\item{log_trans}{logical, should y-axis be log? Defaults to \code{FALSE}}

\item{converted}{logical, were the units converted from the original units used by CDMO? Defaults to \code{FALSE}. See \code{y_labeler} for details.}

\item{plot_title}{logical, should the station name be included as the plot title? Defaults to \code{FALSE}}

\item{plot}{logical, should a plot be returned? Defaults to \code{TRUE}}
}
\value{
A \code{\link[ggplot2]{ggplot}} object
}
\description{
Observed data compared against user-defined percentiles
}
\details{
This function provides an alternative to \code{threshold_plot}. For parameters that may not have numeric threshold criteria, a percentile threshold can be used instead. For a one-tailed analysis, the 90-th percentile is recommended. For a two-tailed analysis, the 5-th and 95-th percentiles are recommended.

Using \code{by_month}, the user can specify whether the percentiles should be calculated on a monthly basis or by using the entire data set.
}
\examples{
\dontrun{
dat_wq <- qaqc(elksmwq, qaqc_keep = c(0, 3, 5))
dat_wq <- subset(dat_wq, subset = '2007-01-01 0:00', operator = '>=')

x <-
  threshold_percentile_plot(dat_wq, param = 'do_mgl'
  , hist_rng = c(2007, 2016), by_month = F)

y <-
  threshold_percentile_plot(dat_wq, param = 'do_mgl'
  , hist_rng = c(2007, 2016), target_yr = 2016, by_month = F)

x2 <-
  threshold_percentile_plot(dat_wq, param = 'do_mgl'
  , hist_rng = c(2007, 2016), by_month = F)

y2 <-
  threshold_percentile_plot(dat_wq, param = 'do_mgl'
  , hist_rng = c(2007, 2016), target_yr = 2016, by_month = F)


dat_nut <- qaqc(elksmnut, qaqc_keep = c(0, 3, 5))
dat_nut <- subset(dat_nut, subset = '2007-01-01 0:00', operator = '>=')
dat_nut <- rem_reps(dat_nut)

x <-
  threshold_percentile_plot(dat_nut, param = 'chla_n'
  , hist_rng = c(2007, 2016), by_month = F)

y <-
  threshold_percentile_plot(dat_nut, param = 'chla_n'
  , hist_rng = c(2007, 2016), target_yr = 2016, by_month = F)
}

}
\seealso{
\code{\link[ggplot2]{ggplot}}, \code{\link{y_labeler}}
}
\author{
Julie Padilla
}
\concept{
analyze

analyze
}
