% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_plot.R
\name{threshold_plot}
\alias{threshold_plot}
\alias{threshold_plot.swmpr}
\title{Seasonal boxplots}
\usage{
threshold_plot(swmpr_in, ...)

\method{threshold_plot}{swmpr}(swmpr_in, param = NULL, hist_rng = NULL,
  target_yr = NULL, thresholds = NULL, log_trans = FALSE, ...)
}
\arguments{
\item{swmpr_in}{input swmpr object}

\item{...}{additional arguments passed to other methods. See \code{\link{assign_season}}}

\item{param}{chr string of variable to plot}

\item{hist_rng}{numeric vector, if historic range is not specified then the min/max values of the data set will be used.}

\item{target_yr}{numeric, the target year that should be compared against the historic range. If target year is not specified then dot will not be plotted}

\item{thresholds}{numeric vector, numeric criteria that will be plotted in the background}

\item{log_trans}{logical, should y-axis be log? Defaults to \code{FALSE}}
}
\value{
A \code{\link[ggplot2]{ggplot}} object
}
\description{
Annual time series for year of interest on top of long-term percentiles
}
\details{
Annual time series for year of interest on top of long-term percentiles
}
\seealso{
\code{\link[ggplot2]{ggplot}}
}
\author{
Julie Padilla
}
\concept{
analyze

analyze
}
