% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_summary.R
\name{threshold_summary}
\alias{threshold_summary}
\alias{threshold_summary.swmpr}
\title{Summary Plots for Threshold Identification}
\usage{
threshold_summary(swmpr_in, ...)

\method{threshold_summary}{swmpr}(swmpr_in, param = NULL,
  summary_type = c("month", "season", "year"), parameter_threshold = NULL,
  threshold_type = NULL, time_threshold = NULL, pal = "Set3",
  plot_title = FALSE, plot = TRUE, ...)
}
\arguments{
\item{swmpr_in}{input swmp object}

\item{...}{additional arguments passed to other methods.}

\item{param}{chr string of variable to plot (one only)}

\item{summary_type}{Choose from \code{month}, \code{season}, or \code{year} aggregation}

\item{parameter_threshold}{vector of numerical thresholds to evaluate parameters against}

\item{threshold_type}{vector of logical operators ('<', '>', '<=', '>=', '==', '!=')}

\item{time_threshold}{The amount of time an event must last to be counted (in hours)}

\item{pal}{Select a palette for boxplot fill colors. See \code{\link[ggplot2]{scale_fill_brewer}} for more details.}

\item{plot_title}{logical, should the station name be included as the plot title? Defaults to \code{FALSE}}

\item{plot}{logical, should a plot be returned? Defaults to \code{TRUE}}
}
\value{
A \code{\link[ggplot2]{ggplot}} object (if \code{plot} = \code{T}) or a dataframe (if \code{plot} = \code{F})
}
\description{
Summary plots for threshold identification analysis
}
\details{
This function produces a plot summary (or a tabular summary) of the results from \code{threshold_identification}.
The user can summarise results on monthly, seasonal, or annual basis by specifying \code{summary_type} = \code{c('month', 'season', 'year')}. The user can specify \code{'month'} for nutrient parameters, but this is not recommended and will produce a warning.
}
\examples{
\dontrun{

## Water quality examples
dat_wq <- qaqc(apacpwq, qaqc_keep = c(0, 3, 5))
dat_wq <- setstep(dat_wq)

x <-
  threshold_summary(dat_wq, param = 'do_mgl', parameter_threshold = 2
  , threshold_type = '<', time_threshold = 2, summary_type = 'month'
  , plot_title = T)

y <-
  threshold_summary(dat_wq, param = 'do_mgl', parameter_threshold = 2,
  threshold_type = '<', time_threshold = 2, summary_type = 'season',
  season = list(c(1,2,3), c(4,5,6), c(7,8,9), c(10, 11, 12)),
  season_names = c('Winter', 'Spring', 'Summer', 'Fall'),
  season_start = 'Winter',
  plot_title = T)


## Nutrient examples
dat_nut <- qaqc(apacpnut, qaqc_keep = c(0, 3, 5))
dat_nut <- setstep(dat_nut, timestep = 'weeks')

x <-
  threshold_summary(dat_nut, param = 'chla_n',
  parameter_threshold = 10,
  threshold_type = '>', time_threshold = 2, summary_type = 'month',
  plot_title = T)

y <-
  threshold_summary(dat_wq, param = 'chla_n', parameter_threshold = 10,
  threshold_type = '>', time_threshold = 2, summary_type = 'season',
  season = list(c(1,2,3), c(4,5,6), c(7,8,9), c(10, 11, 12)),
  season_names = c('Winter', 'Spring', 'Summer', 'Fall'),
  season_start = 'Winter', plot_title = T)

 z <-
   threshold_summary(dat_nut, param = 'chla_n', parameter_threshold = 10,
   threshold_type = '>', time_threshold = 2, summary_type = 'year',
   plot_title = T, plot = T)
}
}
\seealso{
, \code{\link{assign_season}}, \code{\link[ggplot2]{ggplot}}, \code{\link{threshold_identification}}, \code{\link[ggplot2]{scale_fill_brewer}}
}
\author{
Julie Padilla
}
\concept{
analyze

analyze
}
